# api/openapi.yml
components:
  schemas:
    APIError:
      properties:
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
      type: object
    InternshipAdminDto:
      properties:
        description:
          type: string
        goals:
          items:
            type: string
          type: array
          uniqueItems: false
        id:
          type: integer
        label:
          type: string
        skills:
          items:
            type: string
          type: array
          uniqueItems: false
        status:
          type: string
        title:
          type: string
      type: object
    InternshipDto:
      properties:
        description:
          type: string
        goals:
          items:
            type: string
          type: array
          uniqueItems: false
        id:
          type: integer
        label:
          type: string
        skills:
          items:
            type: string
          type: array
          uniqueItems: false
        title:
          type: string
      type: object
    InternshipRequestDto:
      properties:
        description:
          type: string
        goals:
          items:
            type: string
          type: array
          uniqueItems: false
        label:
          type: string
        skills:
          items:
            type: string
          type: array
          uniqueItems: false
        title:
          type: string
      type: object
    InternshipStatusRequestDto:
      properties:
        status:
          type: string
      type: object
  securitySchemes:
    Bearer:
      description: 'Формат: `Bearer jwt_token`'
      in: header
      name: Authorization
      type: apiKey
externalDocs:
  description: ""
  url: ""
info:
  title: Internships API
  version: ""
openapi: 3.1.0
paths:
  /admin/internships:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InternshipAdminDto'
                type: array
          description: OK
      security:
      - Bearer: []
      summary: Получить все стажировки
      tags:
      - admin
    post:
      description: Стажировка создаётся со статусом DRAFT
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternshipRequestDto'
              description: Стажировка
              summary: internship
        description: Стажировка
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipAdminDto'
          description: Created
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unauthorized
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
      security:
      - Bearer: []
      summary: Создать стажировку
      tags:
      - admin
  /admin/internships/{id}:
    get:
      parameters:
      - description: ID стажировки
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipAdminDto'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
      security:
      - Bearer: []
      summary: Получить стажировку по ID
      tags:
      - admin
    patch:
      parameters:
      - description: ID стажировки
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternshipRequestDto'
              description: Стажировка
              summary: internship
        description: Стажировка
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipAdminDto'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
      security:
      - Bearer: []
      summary: Обновить стажировку
      tags:
      - admin
    put:
      parameters:
      - description: ID стажировки
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternshipRequestDto'
              description: Стажировка
              summary: internship
        description: Стажировка
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipAdminDto'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
      security:
      - Bearer: []
      summary: Обновить стажировку
      tags:
      - admin
  /admin/internships/{id}/status:
    put:
      parameters:
      - description: ID стажировки
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternshipStatusRequestDto'
              description: Стажировка
              summary: internship
        description: Стажировка
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipAdminDto'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
      security:
      - Bearer: []
      summary: Обновить статус стажировки
      tags:
      - admin
  /internships:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InternshipDto'
                type: array
          description: OK
      summary: Получить все стажировки
      tags:
      - internship
  /internships/{id}:
    get:
      parameters:
      - description: ID стажировки
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipDto'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
      summary: Получить стажировку по ID
      tags:
      - internship
servers:
- url: /api/v1